Move `fetchLoginMessages` to its own file
